pyproject.toml
[tool.poetry]
name = "todoguardian"
version = "0.1.4"
description = "A python powered web application for handling all your todos in a central location."
authors = ["Bernard Siebens <bernard@siebens.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
python-dateutil = "^2.9.0"
pytodotxt = "^1.5.0"
django-environ = "^0.11.2"
django-compressor = "^4.4"
teamcity-messages = "^1.32"
django-initials-avatar = "^0.0.7"

[tool.poetry.dev-dependencies]
black = "^24.4.2"
flake8-django = "^1.4"
pytest-django = "^4.8.0"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
